Player fires bullets from the pool
The player object gets the pool of bullets when it spawns. When firing,
it will try to find an inactive bullet in the object pool and 'fire'
(reactivate) it at the player's current position. Changed code in other
scripts to no longer destroy bullet objects but to just set them as
inactive.