Bug fixes
Enemies will now switch to No Target mode early if the barrel is somehow
destroyed before the enemy reaches its escape point (e.g. barrel was
already close to exiting the Destroy Area and an enemy nudged it out
when picking it up).
Points are only awarded if there are any barrels left on the screen
(fixed null pointer exception).
Player spawn point changed to be fully within the screen (caused
problems with screen wrapping when spawned right on the edge of the
Destroy Area).